Move types into separate module
