+++ superdesk/core/web/__init__.py
@@ -8,15 +8,14 @@
 # AUTHORS and LICENSE files distributed with this source code, or
 # at https://www.sourcefabric.org/superdesk/license
 
-from .types import (
-    HTTP_METHOD,
-    Response,
-    EndpointFunction,
-    Endpoint,
-    Request,
-    EndpointGroup,
-    RestResponseMeta,
-    RestGetResponse,
-    endpoint,
-    WSGIApp,
-)
+from .endpoints import Endpoint, EndpointGroup, NullEndpoint, endpoint
+from .rest_endpoints import RestEndpoints, ItemRequestViewArgs
+
+__all__ = [
+    "Endpoint",
+    "EndpointGroup",
+    "NullEndpoint",
+    "endpoint",
+    "RestEndpoints",
+    "ItemRequestViewArgs",
+]
